backend/pyproject.toml
[tool.poetry]
name = "Intelligent-Code-Completion-Tool"
version = "0.1.0"
description = "The Intelligent Code Completion Tool provides real-time, context-aware code suggestions and completions based on the context of the code being written, leveraging AI models trained on vast code repositories across multiple programming languages, helping developers write efficient code more quickly and accurately."
authors = ["xxx <xxx@example.com>"]
readme = "../README.md"

packages = [{ include = "../backend" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fuzzywuzzy = "^0.18.0"
Flask-Cors = "^5.0.0"
flask_caching = "^2.3.0"
nest-asyncio = "^1.6.0"
nltk = "^3.9.1"
ollama = "^0.4.5"
psycopg2-binary = "^2.9.10"
python-Levenshtein = "^0.26.1"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.37"
scikit-learn = "^1.6.0"
langchain-groq = "^0.2.2"
langchain-chroma = "^0.1.4"
langchain-community = "^0.3.14"
langchain-text-splitters = "^0.3.4"
spacy = "^3.8.3"
flask = "^3.1.0"
sentence-transformers = "^3.3.1"
cryptography = "^44.0.0"
pycryptodome = "^3.21.0"
apscheduler = "^3.11.0"
pygithub = "^2.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^24.10.0"

[tool.poetry.urls]
repository = "git@github.com:positsource/AI-Code-generator-tool.git"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
